projects pulled from database for all pages
