+++ app/routes/dashboard.py
@@ -1,28 +1,9 @@
 from flask import Blueprint, render_template, session, redirect, request, url_for
 from app.src.database import db
 from app.tables.projects import Project
+from app.src.util_functions import get_all_projects
 
 DashBP= Blueprint('dashboard', __name__)
-
-# temporary project id:name
-# first one should always be add project
-# projects = {'_add_button': 'Add Project', 'alpha': 'Project Alpha',
-#              'beta': 'Project Beta', 'gamma': 'Project Gamma'}
-
-def get_all_projects():
-    '''
-    Retrieve all projects from db
-    returns a dictionary where keys are project ids, and each value holds the title and description
-    '''
-
-    projects = db.session.query(Project).all()
-
-    result = {}
-    for project in projects:
-        result[project.id] = {'description': project.description, 'title': project.title}
-
-    return result
-
 
 @DashBP.before_request
 def require_login():
